feat(server): T-M0.5 真实客户端 e2e PASS（kafka-python）
- e2e：100 条 produce→fetch 往返，不丢/不重/分区有序/offsets_for_times 全过
- 对拍抓出三个协议 bug 并修复：
  1) legacy string 长度前缀是 int16（非 int32；bytes/records 才是 int32）
  2) legacy null string 同为 int16 -1（Null 分支漏改）
  3) batchLength 语义：total = 12 + batchLength（非 61 + batchLength），
     batchLength 从偏移 12 度量、含 leaderEpoch..records
- KIP-511：协商前 ApiVersions 探测（旧头/无 body）失败时回 v0 UNSUPPORTED_VERSION，连接保活
- Metadata 版本语义：AllowAutoTopicCreation 仅 v4+；v1-v3 按 broker 默认允许自动建题
- 协议回归测试：string i16 前缀、版本门控字节数（v1/v7 差 14B）

+++ protocol/src/frame.rs
@@ -43,10 +43,10 @@ pub fn read_request_header(src: &Bytes, api_flexible_from: Option<i16>) -> Resul
         }
     };
 
-    // client_id：legacy string（i32 len），v2 头仍是 legacy 编码（头的 flexible 只体现在 tag section）
+    // client_id：nullable string（int16 长度；头永不使用 compact 编码）
     let mut client_id = None;
     if header_version >= 1 {
-        let len = r.i32()?;
+        let len = r.i16()?;
         if len >= 0 {
             let s = r.take(len as usize)?;
             client_id = Some(String::from_utf8_lossy(s).into());
